{EQEmu PR 2222} [INT64] Further int64 cleanup in Perl SetHP() and GetSpellHPBonuses() in Perl/Lua.
Setter was using int, causing any value over 2 billion to overflow.
Stat bonuses HP is int64, so getter for Spell HP Bonuses should be, too.

+++ zone/lua_mob.h
@@ -123,7 +123,7 @@ public:
 	int64 GetHP();
 	int64 GetMaxHP();
 	int64 GetItemHPBonuses();
-	int GetSpellHPBonuses();
+	int64 GetSpellHPBonuses();
 	double GetWalkspeed();
 	double GetRunspeed();
 	int GetCasterLevel(int spell_id);
